AI for Code Explaination/requirements.txt
langchain
ipykernel
google-generativeai
streamlit
langchain_google_genai
python-dotenv